src/pipeline/requirements.txt
anyio==4.9.0
boto3==1.37.28
botocore==1.37.28
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-pubsub==2.29.0
google-cloud-logging==3.11.4
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jmespath==1.0.1
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.1
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
